refactor: 기존 fitness 패키지 분리
- 테스트에 대한 fitness 패키지 분리
