+++ src/test/java/com/flab/eattofit/exercise/domain/fitness/fitness/FitnessTest.java
@@ -1,6 +1,5 @@
-package com.flab.eattofit.exercise.domain.fitness;
+package com.flab.eattofit.exercise.domain.fitness.fitness;
 
-import com.flab.eattofit.exercise.domain.fitness.fitness.Fitness;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
