+++ src/test/java/com/flab/eattofit/exercise/infrastructure/fitness/fitness/FitnessFakeRepository.java
@@ -1,4 +1,4 @@
-package com.flab.eattofit.exercise.infrastructure.fitness;
+package com.flab.eattofit.exercise.infrastructure.fitness.fitness;
 
 import com.flab.eattofit.exercise.domain.fitness.fitness.Fitness;
 import com.flab.eattofit.exercise.domain.fitness.fitness.FitnessRepository;
